FIXED GOING HOME BUG
FINALLY!!!

Here's what happened. Whenever it moved, it was checking the status. When it had low fuel or too much shit, it would send a call to go home. While it was going home, it was also creating another call to go home every step! So then after it got home, it tried to do all of the resume calls.